+++ CHANGES.md
@@ -23,6 +23,7 @@ Changes in 3.3
 - Check if Request exists before inserting in Requests.php
 - Check for PHP gd extension in diagnostic.php & INSTALL.md
 - Add $FileUploadsPath & $PNGQuantPath optional configuration variables in INSTALL.md & Warehouse.php
+- Add Image resize and compress class in classes/ImageResizeGD.php
 
 Changes in 3.2
 --------------